move decoder implementation in their own files
